Avoid mutable descriptors in modify_range() callbacks
In order to be able to keep track whether or not descriptors are being
modified by the callbacks provided to the modify_range() API, tweak the
prototype so that they return the new descriptor, leaving it to the
library to update the live descriptor.

This is needed for proper TLB maintenance, which is generally only
needed when descriptors change value.