Update modify_flags() to return an error rather than panic!()
Tighten modify_flags() to only permit modifications that are
allowed by the architecture when the descriptor is live. Instead of
panicking when this condition is violated, update the return type and
return a Err() in that case, and leave it to the caller to decide
whether or not panicking is a reasonable course of action.

Wrap this functionality in a helper function apply_masks() that does not
modify the underlying descriptor directly, but returns the updated value
the descriptor should assume as a result of the modification. This will
be used in subsequent patches to phase out the use of mutable live
descriptors in the public API.