Reimplement BBM using apply_masks() and avoid Descriptor temporaries
Now that Descriptors are used to model actual live descriptor in memory
which may be accessed by the page table walker concurrently, we should
avoid using Descriptors where we are only interested in reasoning about
its value. So update the BBM verification routines to use DescriptorBits
instead where the type in question is not an entry in an actual page
table.

Then, re-implement the BBM verification in terms of apply_masks(), which
is a stricter version of the BBM checks, and therefore suitable as a
replacement. Note that this requires setting the TABLE_OR_PAGE flag at
level 3 on the artificial DescriptorBits values constructed on behalf of
map_range().